#96 Create GlobalTestEnvironment and update the code for AbstractAgent.java
